Enable dynamic definition of the embedded app layout
When running a Shopify app in embedded mode, every controller that
requires an authenticated session will use the embedded app layout
by default.

As a default, this is great. It helps the developer to make less
choices when they start to develop their new Shopify app. However,
sometimes one wants to run an embedded app in non-embedded mode.

Some examples of running a normally embedded app in non-embedded mode:
- When the remote environment is a CI;
- When the remote environment is a preview/PR app;
- When the developer wants to run the app as a non-embedded mode for testing.

The rather static implementation of the embedded app layout causes
duplication and maintainability issues when running in both modes.

This commit introduces a more flexible system whereby the deverloper
has more control over what layout is being rendered during the process
and it allows the developer to ignore the embedded app layout.

Instead, the developer can implement the embedded app layout (as
provided by the gem) once in the application.html.erb, and set the

Now, when running the app both in embedded and non-embedded mode,
the developer only has to maintain a single layout file for all
controllers in its app.